fix: :recycle: ts-migration type fixes and refactor
