+++ packages/cfd/src/Components/Routes/binary-routes.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Switch } from 'react-router-dom';
 import { Localize } from '@deriv/translations';
 import getRoutesConfig from '../../Constants/routes-config';
-import { TBinaryRoutes } from '../../types/common-prop-types';
+import { TBinaryRoutes } from '../props.types';
 import RouteWithSubRoutes from './route-with-sub-routes';
 
 const Loading = () => (
