+++ packages/cfd/src/Components/Routes/helpers.ts
@@ -1,6 +1,6 @@
 import { matchPath } from 'react-router';
 import { routes } from '@deriv/shared';
-import { TRouteConfig } from '../../types/common-prop-types';
+import { TRouteConfig } from '../props.types';
 
 export const normalizePath = (path: string) => (/^\//.test(path) ? path : `/${path || ''}`); // Default to '/'
 
@@ -8,16 +8,16 @@ export const findRouteByPath = (path: string, routes_config: TRouteConfig[]) => 
     let result: TRouteConfig | undefined;
 
     routes_config.some(route_info => {
-        let match_path;
+        let has_path_matched;
         try {
-            match_path = matchPath(path, route_info);
+            has_path_matched = matchPath(path, route_info);
         } catch (e) {
             if (/undefined/.test((e as Error).message)) {
                 return undefined;
             }
         }
 
-        if (match_path) {
+        if (has_path_matched) {
             result = route_info;
             return true;
         } else if (route_info.routes) {
