+++ packages/cfd/src/Components/props.types.ts
@@ -1,3 +1,4 @@
+import { Redirect } from 'react-router-dom';
 import { DetailsOfEachMT5Loginid } from '@deriv/api-types';
 import { TCFDPasswordReset } from '../Containers/props.types';
 
@@ -227,4 +228,24 @@ export type TTradingPlatformAccounts = {
      * Name of trading platform.
      */
     platform?: 'dxtrade' | string;
+};
+
+export type TRoute = {
+    exact?: boolean;
+    path: string;
+    default?: boolean;
+    to?: string;
+    component?: ((props?: any) => JSX.Element) | typeof Redirect | React.LazyExoticComponent<() => JSX.Element>;
+    getTitle?: () => string;
+    subroutes?: TRoute[];
+};
+
+export type TRouteConfig = TRoute & {
+    is_authenticated?: boolean;
+    routes?: TRoute[];
+};
+
+export type TBinaryRoutes = {
+    is_logged_in: boolean;
+    is_logging_in: boolean;
 };
