+++ packages/cfd/src/Constants/routes-config.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { routes } from '@deriv/shared';
+import { routes, CFD_PLATFORMS } from '@deriv/shared';
 import { localize } from '@deriv/translations';
 import CFD from '../Containers';
-import { TCFDDashboardProps } from '../Containers/cfd-dashboard';
-import { TRouteConfig } from '../types/common-prop-types';
+import { TRouteConfig } from '../Components/props.types';
 
 // Error Routes
 const Page404 = React.lazy(() => import(/* webpackChunkName: "404" */ '../Modules/Page404'));
@@ -13,13 +12,13 @@ const initRoutesConfig = (): TRouteConfig[] => {
     return [
         {
             path: routes.dxtrade,
-            component: (props: TCFDDashboardProps) => <CFD {...props} platform='dxtrade' />,
+            component: (props: React.ComponentProps<typeof CFD>) => <CFD {...props} platform={CFD_PLATFORMS.DXTRADE} />,
             getTitle: () => localize('Deriv X'),
             is_authenticated: false,
         },
         {
             path: routes.mt5,
-            component: (props: TCFDDashboardProps) => <CFD {...props} platform='mt5' />,
+            component: (props: React.ComponentProps<typeof CFD>) => <CFD {...props} platform={CFD_PLATFORMS.MT5} />,
             getTitle: () => localize('MT5'),
             is_authenticated: false,
         },
